+++ docs/docs/package-manager.md
@@ -6,6 +6,7 @@
     validate, list, install, add, upgrade, remove, purge, clean
 
 With this package manager dependencies can be installed.
+Git is required to be installed on the system.
 
 ## Dependency File
 
@@ -18,7 +19,11 @@ org.alien4cloud.agentpuppet@3.0.x: https://github.com/alien4cloud/csar-public-li
 
 ## Install
 
-Installs all dependencies listed in the `dependencies.yaml` file:
+!!! Info
+    It is recommended to add `lib` to your `.gitignore`.
+
+
+Installs all dependencies listed in the `dependencies.yaml` file in a `lib` directory.
 Already installed dependencies using a branch as checkout will be upgraded instead.
 
 ```shell linenums="1"
@@ -49,5 +54,5 @@ A correctly configured authentication for Git is assumed at this point and is no
 
 ```yaml linenums="1"
 imports:
-- lib/org.alien4cloud.agentpuppet@3.0.x
+- lib/org.alien4cloud.agentpuppet@3.0.x/types.yaml
 ```
