+++ src/controller/setup/clean.ts
@@ -1,5 +1,5 @@
 import * as files from '../../utils/files'
-import config from '../../cli/config'
+import config from '../../utils/config'
 
 export default async function () {
     files.removeDirectory(config.home)
