+++ src/controller/setup/init.ts
@@ -1,5 +1,5 @@
 import * as files from '../../utils/files'
-import config from '../../cli/config'
+import config from '../../utils/config'
 import {Templates} from '../../repository/templates'
 import {Instances} from '../../repository/instances'
 import {Orchestrators} from '../../repository/orchestrators'
