+++ src/controller/setup/open-home.ts
@@ -1,5 +1,5 @@
 import open from 'open'
-import config from '../../cli/config'
+import config from '../../utils/config'
 
 export default async function () {
     await open(config.home)
